fix broken server count
